flask_app/requirements.txt
flask==2.0.1
Werkzeug==2.0.3
flask-cors==3.0.10
paho-mqtt==1.5.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
requests==2.26.0

# Testing
pytest==7.4.0
pytest-mock==3.11.1
